backend/requirements.txt
Django==6.0
djangorestframework==3.17.1
django-cors-headers==4.9.0
pillow==12.2.0

